📂Class_Assignment/openai-agentsdk-assignments/assignment_06/pyproject.toml
[project]
name = "assignment-06"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "load-dotenv>=0.1.0",
    "openai-agents>=0.2.6",
    "pydantic>=2.11.7",
]
